requirements.txt
requests==2.31.0
pytz==2023.3
google-auth==2.22.0
google-api-python-client==2.93.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
numpy>=1.24.0
pandas>=2.0.0 
seaborn>=0.13.0 
plotly>=5.18.0 